// frontend/src/app/api/model/userRegistrationRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleDto } from './roleDto';


export interface UserRegistrationRequestDto { 
    name?: string;
    email?: string;
    password?: string;
    roleDto?: RoleDto;
}